// src/interfaces/CustomerData.ts
export interface CustomerData {
  fullName: string
  email: string
  mobile: string
  document: string
  zipCode: string
  address: string
  number: string
  complement: string
  neighborhood: string
  city: string
  state: string
  creditCardNumber: string
  creditCardHolderName: string
  creditCardExpirationDate: string
  creditCardCVV: string
}